UI 수정

+++ app/src/main/java/finotek/global/dev/talkbank_ca/chat/Scenario.java
@@ -49,7 +49,7 @@ class Scenario {
         RxEventBus.getInstance().getObservable()
             .subscribe(iEvent -> {
                 if (iEvent instanceof AccuracyMeasureEvent) {
-                    double accuracy = ((AccuracyMeasureEvent) iEvent).getAccuracy();
+                    int accuracy = ((AccuracyMeasureEvent) iEvent).getAccuracy();
                     messageBox.add(new StatusMessage("맥락 데이터 분석 결과 " + String.valueOf(accuracy * 100) + "% 확률로 인증되었습니다."));
                 }
             });
